Add So ability (previous commit was Kho ability)

+++ characters/Baileys/Blonde.java
@@ -11,7 +11,7 @@ public class Blonde extends Bailey {
     public static final String IMAGE_NAME = "blonde.png";
     static final int startingSpeed = 10;
     static final int d = 100;
-    static final int r = 1;
+    static final double r = 1;
     public Blonde(int x, int y, int w, int h) {
         super(name, startingHp, IMAGE_NAME, x, y, w, h, startingSpeed, d, r);
     }
